src/requirements.txt
objectdaddy==0.0.4
rtsparty==0.0.1
RPi.GPIO==0.7.1
flask
requests
